// Frontend/SistemaFlota/src/app/core/models/compras-no-formalizadas/proveedores/proveedor-no-formalizado.model.ts
export interface ProveedorNoFormalizado {
    idProveedorNoFormalizado?: number;
    nombre: string;
    documento?: string;
    contacto?: string;
    telefono?: string;
    correoElectronico?: string;
    direccion?: string;
    ciudad?: string;
    departamento?: string;
    activo?: boolean;
    fechaCreacion?: string;
    fechaActualizacion?: string;
}